refactor default test helpers
