+++ solarforecastarbiter/conftest.py
@@ -114,6 +114,48 @@ def ghi_observation_metadata(site_metadata):
     return ghi_meta
 
 
+def default_observation(
+        site_metadata,
+        name='Albuquerque Baseline', variable='ghi',
+        value_type='mean', uncertainty=1, interval_length='1h',
+        interval_label='beginning'):
+    """
+    Helpful when you want to test a couple of parameters and don't
+    need to worry about the rest.
+    """
+    obs = datamodel.Observation(
+        name=name, variable=variable,
+        value_type=value_type, site=site_metadata, uncertainty=uncertainty,
+        interval_length=pd.Timedelta(interval_length),
+        interval_label=interval_label
+    )
+    return obs
+
+
+def default_forecast(
+        site_metadata,
+        name='Albuquerque Baseline', variable='ghi',
+        value_type='mean', issue_time_of_day=dt.time(hour=5),
+        lead_time_to_start='1h', interval_length='1h', run_length='12h',
+        interval_label='beginning'):
+    """
+    Helpful when you want to test a couple of parameters and don't
+    need to worry about the rest.
+    """
+    fx = datamodel.Forecast(
+        site=site_metadata,
+        name=name,
+        value_type=value_type,
+        variable=variable,
+        issue_time_of_day=issue_time_of_day,
+        lead_time_to_start=pd.Timedelta(lead_time_to_start),
+        interval_length=pd.Timedelta(interval_length),
+        run_length=pd.Timedelta(run_length),
+        interval_label=interval_label
+    )
+    return fx
+
+
 @pytest.fixture(scope='module')
 def ac_power_forecast_metadata(site_metadata):
     ac_power_fx_meta = datamodel.Forecast(
